# backend/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: POS Platform API
  version: 0.1.0
  description: >
    REST API for the POS platform providing checkout, backoffice, and
    synchronization capabilities. The API is designed to be idempotent, secure,
    and ready for hybrid deployments.
servers:
  - url: http://localhost:8080
    description: Local development
...
security:
  - bearerAuth: []
paths:
...
  /api/v1/sellers:
    get:
      tags:
        - Catalog
      summary: List sellers
      parameters:
        - $ref: '#/components/parameters/SearchQuery'
        - $ref: '#/components/parameters/LimitQuery'
      responses:
        '200':
          description: Sellers ordered by name.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Seller'
    post:
      tags:
        - Catalog
      summary: Create seller
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SellerWrite'
      responses:
        '201':
          description: Seller created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Seller'
        '400':
          $ref: '#/components/responses/ValidationError'
        '409':
          $ref: '#/components/responses/ConflictError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
  /api/v1/sellers/{id}:
    get:
      tags:
        - Catalog
      summary: Get seller
      parameters:
        - $ref: '#/components/parameters/ResourceId'
      responses:
        '200':
          description: Seller.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Seller'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
    patch:
      tags:
        - Catalog
      summary: Update seller
      parameters:
        - $ref: '#/components/parameters/ResourceId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SellerPatch'
      responses:
        '200':
          description: Seller updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Seller'
        '204':
          description: No fields changed.
        '400':
          $ref: '#/components/responses/ValidationError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
...
  /api/v1/sales:
    get:
      tags:
        - Sales
      summary: List sales
      parameters:
        - name: from
          in: query
          description: Filter by emission date (inclusive, YYYY-MM-DD).
          schema:
            type: string
            format: date
        - name: to
          in: query
          description: Filter by emission date (inclusive, YYYY-MM-DD).
          schema:
            type: string
            format: date
        - name: seller_id
          in: query
          description: Filter by seller identifier.
          schema:
            type: string
            format: uuid
        - name: customer_id
          in: query
          description: Filter by customer identifier.
          schema:
            type: string
            format: uuid
        - name: payment_term_id
          in: query
          description: Filter by payment term identifier.
          schema:
            type: string
            format: uuid
        - name: limit
          in: query
          description: Maximum number of records to return (default 100, max 200).
          schema:
            type: integer
            minimum: 1
            maximum: 200
      responses:
        '200':
          description: Most recent sales first.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Sale'
    post:
      tags:
        - Sales
      summary: Register sale
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SaleCreate'
      responses:
        '201':
          description: Sale created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Sale'
        '400':
          $ref: '#/components/responses/ValidationError'
        '409':
          $ref: '#/components/responses/ConflictError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
  /api/v1/sales/{id}:
    get:
      tags:
        - Sales
      summary: Get sale
      parameters:
        - $ref: '#/components/parameters/ResourceId'
      responses:
        '200':
          description: Sale with items.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Sale'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
  /api/v1/sales/{id}/cancel:
    post:
      tags:
        - Sales
      summary: Cancel sale
      description: >
        Cancels a sale when permitted by RBAC policy. Previously cancelled sales
        return 404 to prevent idempotency violations.
      parameters:
        - $ref: '#/components/parameters/ResourceId'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SaleCancel'
      responses:
        '200':
          description: Sale cancelled.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Sale'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
...
components:
  parameters:
    ResourceId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    SearchQuery:
      name: search
      in: query
      description: Case insensitive partial match on name, legacy code, or identifier.
      schema:
        type: string
    LimitQuery:
      name: limit
      in: query
      description: Maximum number of records to return (default 100, max 500).
      schema:
        type: integer
        minimum: 1
        maximum: 500
  responses:
    ValidationError:
      description: Request payload failed validation.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ValidationError'
    NotFoundError:
      description: Resource not found.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            default:
              value:
                message: resource not found
    ConflictError:
      description: Unique constraint violation.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            default:
              value:
                message: conflict
    UnauthorizedError:
      description: Authentication credentials are missing or invalid.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            default:
              value:
                message: unauthorized
    ForbiddenError:
      description: Authenticated user lacks permission for the requested action.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            default:
              value:
                message: forbidden
  schemas:
...
    Seller:
      type: object
      properties:
        id:
          type: string
          format: uuid
        legacy_code:
          type: string
          nullable: true
        name:
          type: string
      required:
        - id
        - name
    SellerWrite:
      type: object
      properties:
        legacy_code:
          type: string
          nullable: true
        name:
          type: string
      required:
        - name
    SellerPatch:
      allOf:
        - $ref: '#/components/schemas/SellerWrite'
...
    SaleItem:
      type: object
      properties:
        id:
          type: string
          format: uuid
        product_id:
          type: string
          format: uuid
        quantity:
          type: number
          format: float
        unit_price:
          type: number
          format: float
        total:
          type: number
          format: float
      required:
        - id
        - product_id
        - quantity
        - unit_price
        - total
    Sale:
      type: object
      properties:
        id:
          type: string
          format: uuid
        emission_date:
          type: string
          format: date
        order_number:
          type: string
          nullable: true
        seller_id:
          type: string
          format: uuid
          nullable: true
        customer_id:
          type: string
          format: uuid
          nullable: true
        payment_term_id:
          type: string
          format: uuid
          nullable: true
        subtotal:
          type: number
          format: float
        discount:
          type: number
          format: float
        total:
          type: number
          format: float
        status:
          type: string
          enum:
            - draft
            - completed
            - cancelled
        source:
          type: string
          nullable: true
        source_key:
          type: string
          nullable: true
        cancelled_at:
          type: string
          format: date-time
          nullable: true
        cancellation_reason:
          type: string
          nullable: true
        items:
          type: array
          items:
            $ref: '#/components/schemas/SaleItem'
      required:
        - id
        - emission_date
        - subtotal
        - discount
        - total
        - status
        - items
    SaleItemWrite:
      type: object
      properties:
        product_id:
          type: string
          format: uuid
        quantity:
          type: number
          format: float
          minimum: 0
        unit_price:
          type: number
          format: float
          minimum: 0
        total:
          type: number
          format: float
          nullable: true
      required:
        - product_id
        - quantity
        - unit_price
    SaleCreate:
      type: object
      properties:
        emission_date:
          type: string
          format: date
        order_number:
          type: string
          nullable: true
        seller_id:
          type: string
          format: uuid
          nullable: true
        customer_id:
          type: string
          format: uuid
          nullable: true
        payment_term_id:
          type: string
          format: uuid
          nullable: true
        subtotal:
          type: number
          format: float
          nullable: true
        discount:
          type: number
          format: float
          nullable: true
        total:
          type: number
          format: float
          nullable: true
        source:
          type: string
          nullable: true
        source_key:
          type: string
          nullable: true
        items:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/SaleItemWrite'
      required:
        - items
    SaleCancel:
      type: object
      properties:
        reason:
          type: string
          nullable: true
...
    ErrorResponse:
      type: object
      properties:
        message:
          type: string
        details:
          nullable: true
      required:
        - message
    ValidationError:
      type: object
      properties:
        message:
          type: string
          example: validation_failed
        issues:
          type: array
          items:
            type: object
            properties:
              path:
                type: array
                items:
                  type: string
              message:
                type: string
              code:
                type: string
            required:
              - path
              - message
              - code
      required:
        - message
        - issues
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT